requirements.txt
sqlalchemy
psycopg2-binary
pandas
numpy
beautifulsoup4
requests
scrapy
python-dotenv
PyMuPDF
pytesseract
Pillow
tqdm
sentence-transformers
pandas
chromadb
scipy
neo4j
selenium
webdriver-manager